29.04.25 mailService 6.8

+++ node/src/main/java/pro/masterfood/service/impl/AppUserServiceImpl.java
@@ -150,12 +150,27 @@ public class AppUserServiceImpl implements AppUserService {
     @Override
     @Transactional // Важно!
     public String sendReportMail(Long chatId, AppUser appUser) {
-        Long userId = appUser.getId();
+
         try {
-            appPhotoDAO.deleteByOwnerId(userId);//
+            Long userId = appUser.getId();
+            AppUser userForSession = appUserDAO.findById(appUser.getId()).orElse(null); // Получаем пользователя для сессии
+            if (userForSession != null) {
+                // 1. Получаем фотографии. Важно, чтобы они были загружены в рамках транзакции.
+                List<AppPhoto> appPhotos = userForSession.getPhotos();
+                // 2. Создаем List<byte[]> для всех вложений (перед удалением, но в рамках той же транзакции)
+                List<byte[]> attachments = new ArrayList<>();
+                for (AppPhoto appPhoto : appPhotos) {
+                    if (appPhoto.getBinaryContent() != null) {
+                        byte[] binaryContent = appPhoto.getBinaryContent().getFileAsArrayOfBytes();
+                        attachments.add(binaryContent);
+                    }
+                }
+                appPhotoDAO.deleteByOwnerId(userId);//
+                return "Все фотографии пользователя с ID " + userId + " удалены.";
 
-            return "Все фотографии пользователя с ID " + userId + " удалены.";
-
+            } else {
+                return "Пользователь не найден в методе sendReportMail";
+            }
         } catch (Exception e) {
             e.printStackTrace();
             return "Ошибка при удалении фотографий: " + e.getMessage();
